refactor(chatbox): convert ChatBox into a pure UI component
