+++ frontend/src/components/chatbox/ChatBox.jsx
@@ -1,40 +1,28 @@
-// frontend\src\components\chatbox\ChatBox.jsx
-import React, { useRef, useState, useEffect } from "react";
+import React, { useRef } from "react";
 import ChatMessage from "./chatMessage";
-import { useUserAnswers } from "../../context/UserAnswersContext";
 import useChatSound from "../../hooks/assistant/useChatSound";
 import useChatAutoScroll from "../../hooks/assistant/useChatAutoScroll";
 import "../../styles/features/chatbox/ChatBox.css";
 
-export default function ChatBox({ messages = [], onSelect, collapse }) {
+export default function ChatBox({
+  messages = [],
+  onSelect,
+  collapsed,
+  onToggleCollapse,
+  levelLabel
+}) {
+
   const chatContainerRef = useRef(null);
-  const [isCollapsed, setIsCollapsed] = useState(false);
-
-  const { level } = useUserAnswers();
-
-  const levelLabelMap = {
-    beginner: "Principiante",
-    intermediate: "Intermedio",
-    expert: "Experto",
-  };
-
-  const levelLabel = level ? levelLabelMap[level] : null;
-
-  useEffect(() => {
-    if (collapse) {
-      setTimeout(() => setIsCollapsed(true), 150);
-    }
-  }, [collapse]);
 
   useChatSound(messages);
-  useChatAutoScroll(messages, chatContainerRef, isCollapsed);
+  useChatAutoScroll(messages, chatContainerRef, collapsed);
 
   return (
     <div className="chatbox-wrapper">
 
       <div
         className="chatbox-header"
-        onClick={() => setIsCollapsed(prev => !prev)}
+        onClick={onToggleCollapse}
       >
         <span className="chatbox-title">Coffe</span>
 
@@ -45,7 +33,7 @@ export default function ChatBox({ messages = [], onSelect, collapse }) {
         )}
 
         <span
-          className={`chatbox-chevron ${isCollapsed ? "collapsed" : ""}`}
+          className={`chatbox-chevron ${collapsed ? "collapsed" : ""}`}
         >
           ▾
         </span>
@@ -53,7 +41,7 @@ export default function ChatBox({ messages = [], onSelect, collapse }) {
 
       <div
         ref={chatContainerRef}
-        className={`chatbox-container ${isCollapsed ? "collapsed" : ""}`}
+        className={`chatbox-container ${collapsed ? "collapsed" : ""}`}
       >
         {messages.map((msg) => (
           <ChatMessage
@@ -63,6 +51,7 @@ export default function ChatBox({ messages = [], onSelect, collapse }) {
           />
         ))}
       </div>
+
     </div>
   );
 }
