+++ frontend/src/hooks/assistant/useChatSound.js
@@ -1,3 +1,4 @@
+// frontend\src\hooks\assistant\useChatSound.js
 import { useEffect, useRef } from "react";
 
 export default function useChatSound(messages) {
